+++ Persistencia/Interfaces/ICrud.cs
@@ -11,7 +11,7 @@ namespace Persistencia.Interfaces
         Task Add(T Objeto);
         Task Update(T Objeto);
         Task Delete(T Objeto);
-        Task<T> GetEntityById(int Id);
+        Task<T> GetEntityById(long Id);
         Task<IEnumerable<T>> List();
     }
 }
